design: SearchBar 컴포넌트 분리 및 SearchTagBar 컴포넌트 UI 구현
