+++ src/components/molecules/SearchTagBar/SearchTagBar.style.ts
@@ -1,0 +1,13 @@
+import { css } from '@emotion/react';
+
+export const searchBarContainerStyling = css({
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  gap: '20px',
+});
+
+export const btnWrapperStyling = css({
+  display: 'flex',
+  gap: '14px',
+});
