+++ src/stories/atoms/SearchBar.stories.tsx
@@ -1,14 +1,14 @@
-import SearchTagBar from '@/components/molecules/SearchTagBar/SearchTagBar';
+import SearchBar from '@/components/atoms/SearchBar/SearchBar';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
-  title: 'molecules/SearchTagBar',
-  component: SearchTagBar,
+  title: 'atoms/SearchBar',
+  component: SearchBar,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof SearchTagBar>;
+} satisfies Meta<typeof SearchBar>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
